requirements.txt
fastapi
poke-battle-sim
gunicorn
uvicorn
python-dotenv
requests
pandas
